requirements.txt
discord.py
Pillow
ImageHash
aiosqlite
aiohttp
scikit-learn==1.6.1
xgboost
joblib
dill